+++ 1.20.1/Common/src/main/java/fuzs/swordblockingmechanics/capability/ParryCooldownCapability.java
@@ -1,15 +1,14 @@
 package fuzs.swordblockingmechanics.capability;
 
 import fuzs.puzzleslib.api.capability.v2.data.CapabilityComponent;
-import net.minecraft.world.entity.player.Player;
 
 public interface ParryCooldownCapability extends CapabilityComponent {
 
     double getCooldownProgress();
 
-    void setCooldownTicks();
+    void resetCooldownTicks();
 
-    void tick(Player player);
+    void tick();
 
     default boolean isCooldownActive() {
         return this.getCooldownProgress() < 1.0;