+++ 1.21.3/Common/src/main/java/fuzs/swordblockingmechanics/client/handler/FirstPersonRenderingHandler.java
@@ -2,44 +2,60 @@ package fuzs.swordblockingmechanics.client.handler;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
-import fuzs.puzzleslib.api.client.event.v1.renderer.RenderHandEvents;
+import fuzs.puzzleslib.api.client.util.v1.RenderPropertyKey;
 import fuzs.puzzleslib.api.event.v1.core.EventResult;
 import fuzs.swordblockingmechanics.SwordBlockingMechanics;
 import fuzs.swordblockingmechanics.config.ClientConfig;
 import fuzs.swordblockingmechanics.handler.SwordBlockingHandler;
 import fuzs.swordblockingmechanics.mixin.client.accessor.ItemInHandRendererAccessor;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.renderer.ItemInHandRenderer;
 import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.entity.state.EntityRenderState;
+import net.minecraft.client.renderer.entity.state.PlayerRenderState;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.HumanoidArm;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
 
 public class FirstPersonRenderingHandler {
+    public static final RenderPropertyKey<Boolean> IS_BLOCKING_RENDER_PROPERTY_KEY = new RenderPropertyKey<>(
+            SwordBlockingMechanics.id("is_blocking"));
 
-    public static RenderHandEvents.MainHand onRenderHand(InteractionHand interactionHand) {
-        return (itemInHandRenderer, player, humanoidArm, itemStack, poseStack, multiBufferSource, combinedLight, partialTick, interpolatedPitch, swingProgress, equipProgress) -> {
-            return onRenderHand(player, interactionHand, itemStack, poseStack, multiBufferSource, combinedLight, partialTick, interpolatedPitch, swingProgress,equipProgress);
-        };
+    public static void onExtractRenderState(Entity entity, EntityRenderState entityRenderState, float partialTick) {
+        if (entity instanceof Player player && entityRenderState instanceof PlayerRenderState) {
+            RenderPropertyKey.setRenderProperty(entityRenderState,
+                    IS_BLOCKING_RENDER_PROPERTY_KEY,
+                    SwordBlockingHandler.isActiveItemStackBlocking(player));
+        }
     }
 
-    public static EventResult onRenderHand(Player player, InteractionHand interactionHand, ItemStack itemStack, PoseStack poseStack, MultiBufferSource multiBufferSource, int combinedLight, float partialTick, float interpolatedPitch, float swingProgress, float equipProgress) {
+    public static EventResult onRenderBothHands(ItemInHandRenderer itemInHandRenderer, InteractionHand interactionHand, AbstractClientPlayer player, HumanoidArm humanoidArm, ItemStack itemStack, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, float partialTick, float interpolatedPitch, float swingProgress, float equipProgress) {
         if (player.getUsedItemHand() == interactionHand && SwordBlockingHandler.isActiveItemStackBlocking(player)) {
-            Minecraft minecraft = Minecraft.getInstance();
-            ItemInHandRenderer itemRenderer = minecraft.getEntityRenderDispatcher().getItemInHandRenderer();
             poseStack.pushPose();
             boolean mainHand = interactionHand == InteractionHand.MAIN_HAND;
             HumanoidArm handSide = mainHand ? player.getMainArm() : player.getMainArm().getOpposite();
             boolean isHandSideRight = handSide == HumanoidArm.RIGHT;
-            ((ItemInHandRendererAccessor) itemRenderer).swordblockingmechanics$callApplyItemArmTransform(poseStack, handSide, equipProgress);
+            ((ItemInHandRendererAccessor) itemInHandRenderer).swordblockingmechanics$callApplyItemArmTransform(poseStack,
+                    handSide,
+                    equipProgress);
             if (SwordBlockingMechanics.CONFIG.get(ClientConfig.class).interactAnimations) {
-                ((ItemInHandRendererAccessor) itemRenderer).swordblockingmechanics$callApplyItemArmAttackTransform(poseStack, handSide, swingProgress);
+                ((ItemInHandRendererAccessor) itemInHandRenderer).swordblockingmechanics$callApplyItemArmAttackTransform(
+                        poseStack,
+                        handSide,
+                        swingProgress);
             }
             transformBlockFirstPerson(poseStack, handSide);
-            itemRenderer.renderItem(player, itemStack, isHandSideRight ? ItemDisplayContext.FIRST_PERSON_RIGHT_HAND : ItemDisplayContext.FIRST_PERSON_LEFT_HAND, !isHandSideRight, poseStack, multiBufferSource, combinedLight);
+            itemInHandRenderer.renderItem(player,
+                    itemStack,
+                    isHandSideRight ? ItemDisplayContext.FIRST_PERSON_RIGHT_HAND :
+                            ItemDisplayContext.FIRST_PERSON_LEFT_HAND,
+                    !isHandSideRight,
+                    poseStack,
+                    bufferSource,
+                    combinedLight);
             poseStack.popPose();
             return EventResult.INTERRUPT;
         } else {