+++ 1.21.3/Common/src/main/java/fuzs/swordblockingmechanics/client/helper/AdvancedBlockingRenderer.java
@@ -2,32 +2,32 @@ package fuzs.swordblockingmechanics.client.helper;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.ArmedModel;
-import net.minecraft.client.renderer.ItemInHandRenderer;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.block.model.ItemTransform;
+import net.minecraft.client.renderer.entity.ItemRenderer;
+import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.world.entity.HumanoidArm;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
 import org.joml.Quaternionf;
 
 public class AdvancedBlockingRenderer {
 
-    public static void renderBlockingWithSword(ItemInHandRenderer itemInHandRenderer, ArmedModel model, LivingEntity entity, ItemStack stack, ItemDisplayContext transform, HumanoidArm arm, PoseStack poseStack, MultiBufferSource multiBufferSource, int combinedLight) {
+    public static void renderBlockingWithSword(ItemRenderer itemRenderer, ArmedModel model, ItemStack itemStack, ItemDisplayContext itemDisplayContext, HumanoidArm humanoidArm, PoseStack poseStack, MultiBufferSource bufferSource, int packedLight, BakedModel bakedModel) {
         // those transformations are directly ported from Minecraft 1.7, resulting in a pixel-perfect recreation of third-person sword blocking
         // a lot has changed since then (the whole model system has been rewritten twice in 1.8 and 1.9, and had major changes in 1.14 and 1.15),
         // so we reset everything vanilla does now, and apply every single step that was done in 1.7
         // (there were multiple classes and layers involved in 1.7, it is noted down below which class every transformation came from)
         // all this is done in code and not using some custom json model predicate so that every item is supported by default
         poseStack.pushPose();
-        model.translateToHand(arm, poseStack);
-        boolean leftHand = arm == HumanoidArm.LEFT;
+        model.translateToHand(humanoidArm, poseStack);
+        boolean leftHand = humanoidArm == HumanoidArm.LEFT;
         applyItemBlockingTransform(poseStack, leftHand);
-        applyItemTransformInverse(entity, stack, transform, poseStack, leftHand);
-        itemInHandRenderer.renderItem(entity, stack, transform, leftHand, poseStack, multiBufferSource, combinedLight);
+        // revert 1.8+ model changes, so we can work on a blank slate
+        applyTransformInverse(bakedModel.getTransforms().getTransform(itemDisplayContext), leftHand, poseStack);
+        itemRenderer.render(itemStack, itemDisplayContext, leftHand, poseStack, bufferSource, packedLight, OverlayTexture.NO_OVERLAY, bakedModel);
         poseStack.popPose();
     }
 
@@ -57,24 +57,17 @@ public class AdvancedBlockingRenderer {
         poseStack.translate(0.0F, 0.0F, 0.28125F);
     }
 
-    private static void applyItemTransformInverse(LivingEntity entity, ItemStack stack, ItemDisplayContext transform, PoseStack poseStack, boolean leftHand) {
-        // revert 1.8+ model changes, so we can work on a blank slate
-        Minecraft minecraft = Minecraft.getInstance();
-        BakedModel model = minecraft.getItemRenderer().getModel(stack, entity.level(), entity, 0);
-        applyTransformInverse(model.getTransforms().getTransform(transform), leftHand, poseStack);
-    }
-
-    private static void applyTransformInverse(ItemTransform vec, boolean leftHand, PoseStack matrixStackIn) {
+    private static void applyTransformInverse(ItemTransform itemTransform, boolean leftHand, PoseStack poseStack) {
         // this does the exact inverse of ItemTransform::apply which should be applied right after, so that in the end nothing has changed
-        if (vec != ItemTransform.NO_TRANSFORM) {
-            float angleX = vec.rotation.x();
-            float angleY = leftHand ? -vec.rotation.y() : vec.rotation.y();
-            float angleZ = leftHand ? -vec.rotation.z() : vec.rotation.z();
+        if (itemTransform != ItemTransform.NO_TRANSFORM) {
+            float angleX = itemTransform.rotation.x();
+            float angleY = leftHand ? -itemTransform.rotation.y() : itemTransform.rotation.y();
+            float angleZ = leftHand ? -itemTransform.rotation.z() : itemTransform.rotation.z();
             Quaternionf quaternion = new Quaternionf().rotationXYZ(angleX * 0.017453292F, angleY * 0.017453292F, angleZ * 0.017453292F);
             quaternion.conjugate();
-            matrixStackIn.scale(1.0F / vec.scale.x(), 1.0F / vec.scale.y(), 1.0F / vec.scale.z());
-            matrixStackIn.mulPose(quaternion);
-            matrixStackIn.translate((leftHand ? -1.0F : 1.0F) * -vec.translation.x(), -vec.translation.y(), -vec.translation.z());
+            poseStack.scale(1.0F / itemTransform.scale.x(), 1.0F / itemTransform.scale.y(), 1.0F / itemTransform.scale.z());
+            poseStack.mulPose(quaternion);
+            poseStack.translate((leftHand ? -1.0F : 1.0F) * -itemTransform.translation.x(), -itemTransform.translation.y(), -itemTransform.translation.z());
         }
     }
 }