+++ 1.21.3/Common/src/main/java/fuzs/swordblockingmechanics/mixin/client/HumanoidModelMixin.java
@@ -1,14 +1,17 @@
 package fuzs.swordblockingmechanics.mixin.client;
 
+import fuzs.puzzleslib.api.client.util.v1.RenderPropertyKey;
 import fuzs.swordblockingmechanics.SwordBlockingMechanics;
+import fuzs.swordblockingmechanics.client.handler.FirstPersonRenderingHandler;
 import fuzs.swordblockingmechanics.config.ClientConfig;
-import fuzs.swordblockingmechanics.handler.SwordBlockingHandler;
-import net.minecraft.client.model.AgeableListModel;
+import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.geom.ModelPart;
+import net.minecraft.client.renderer.entity.state.HumanoidRenderState;
+import net.minecraft.client.renderer.entity.state.PlayerRenderState;
+import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.HumanoidArm;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -16,25 +19,37 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(HumanoidModel.class)
-abstract class HumanoidModelMixin<T extends LivingEntity> extends AgeableListModel<T> {
+abstract class HumanoidModelMixin<T extends HumanoidRenderState> extends EntityModel<T> {
     @Shadow
     public ModelPart rightArm;
     @Shadow
     public ModelPart leftArm;
 
-    @Inject(method = "setupAnim", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/model/HumanoidModel;setupAttackAnimation(Lnet/minecraft/world/entity/LivingEntity;F)V"))
-    public void setupAnim(T entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, CallbackInfo callback) {
-        if (entity instanceof Player player) {
-            if (SwordBlockingHandler.isActiveItemStackBlocking(player)) {
-                if (entity.getUsedItemHand() == InteractionHand.OFF_HAND) {
-                    this.leftArm.xRot = this.leftArm.xRot - ((float) Math.PI * 2.0F) / 10.0F;
+    protected HumanoidModelMixin(ModelPart root) {
+        super(root);
+    }
+
+    @Inject(
+            method = "setupAnim", at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/client/model/HumanoidModel;setupAttackAnimation(Lnet/minecraft/client/renderer/entity/state/HumanoidRenderState;F)V"
+    )
+    )
+    public void setupAnim(T renderState, CallbackInfo callback) {
+        if (renderState instanceof PlayerRenderState playerRenderState && renderState.isUsingItem) {
+            if (RenderPropertyKey.getRenderProperty(renderState,
+                    FirstPersonRenderingHandler.IS_BLOCKING_RENDER_PROPERTY_KEY)) {
+                InteractionHand interactionHand =
+                        renderState.mainArm == HumanoidArm.RIGHT ? InteractionHand.MAIN_HAND : InteractionHand.OFF_HAND;
+                if (renderState.useItemHand == interactionHand) {
+                    this.rightArm.xRot = this.rightArm.xRot - Mth.PI * 2.0F / 10.0F;
                     if (SwordBlockingMechanics.CONFIG.get(ClientConfig.class).simpleBlockingPose) {
-                        this.leftArm.yRot = ((float) Math.PI / 6.0F);
+                        this.rightArm.yRot = -Mth.PI / 6.0F;
                     }
                 } else {
-                    this.rightArm.xRot = this.rightArm.xRot - ((float) Math.PI * 2.0F) / 10.0F;
+                    this.leftArm.xRot = this.leftArm.xRot - Mth.PI * 2.0F / 10.0F;
                     if (SwordBlockingMechanics.CONFIG.get(ClientConfig.class).simpleBlockingPose) {
-                        this.rightArm.yRot = ((float) -Math.PI / 6.0F);
+                        this.leftArm.yRot = Mth.PI / 6.0F;
                     }
                 }
             }